refactor: serverinitialization script finished
