+++ scripts/server/reset_rabbitmq.py
@@ -1,4 +1,4 @@
-from run_server import rabbitmq_reset_and_shut_down_server
+from utils._rabbitmq import rabbitmq_reset_and_shut_down_server
 
 
 if __name__ == '__main__':
