+++ scripts/server/utils/_celery.py
@@ -1,0 +1,5 @@
+from subprocess import Popen
+
+
+def start_celery_worker(module: str):
+    return Popen(['celery', '-A', module, 'worker'])
